feat: 수동 및 자동 로또 티켓의 개수를 나타내는 TicketCounts 객체 생성
