+++ src/test/java/lotto/domain/TicketCountsTest.java
@@ -6,12 +6,12 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class TicketCountsValidatorTest {
+class TicketCountsTest {
     @DisplayName("수동으로 구매할 로또 티켓의 수가 전체 로또 티켓의 수보다 클 경우 예외 발생")
     @Test
     void validateCountIsGreaterThanMax() {
         assertThatThrownBy(() -> {
-            TicketCountsValidator.validateManualTicketCount(14, 15);
+            TicketCounts.fromMoneyAndManualTicketCount(Money.ofPurchaseMoney(14000), 15);
         }).isInstanceOf(TicketCountsException.class)
                 .hasMessage("구매 가능한 로또 티켓의 수를 초과하였습니다.");
     }
@@ -20,7 +20,7 @@ class TicketCountsValidatorTest {
     @Test
     void validateCountIsGreaterThanMin() {
         assertThatThrownBy(() -> {
-            TicketCountsValidator.validateManualTicketCount(14, 0);
+            TicketCounts.fromMoneyAndManualTicketCount(Money.ofPurchaseMoney(14000), 0);
         }).isInstanceOf(TicketCountsException.class)
                 .hasMessage("구매할 로또 티켓의 수는 양수이어야 합니다.");
     }
